requirements.txt
# General dependencies
requests
jsonschema
python-dotenv
Groq
langchain    
CrewAI
apify

# Optional dependencies
pandas       
numpy
